requirements.txt
anyio==3.6.2
appdirs==1.4.4
argcomplete==2.0.0
click==8.1.3
Deprecated==1.2.13
fastapi==0.89.1
h11==0.14.0
idna==3.4
limits==2.8.0
packaging==22.0
Pillow==9.4.0
pydantic==1.10.4
pypng==0.20220715.0
pyserial==3.5
python-escpos==2.2.0
pyusb==1.2.1
PyYAML==6.0
qrcode==7.4
six==1.16.0
slowapi==0.1.7
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
wrapt==1.14.1
